refactor: nest reader predicates
